requirements.txt
cloudpickle==1.6.0
joblib==1.3.1
numpy==1.24.3
Platypus-Opt==1.0.4
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==2.2.0
